BE/requirements.txt
fastapi 
uvicorn
python-multipart
langchain_community
langchain_huggingface
pymupdf
langchain-ollama